Split ApiMethodTransformer
